% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ask_harvest_account}
\alias{ask_harvest_account}
\title{Ask user for Harvest account ID}
\usage{
ask_harvest_account(msg = "Harvest API Account ID")
}
\arguments{
\item{msg}{message in the Rstudio prompt}
}
\description{
Ask user for Harvest account ID
}
