% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ask_harvest_pat}
\alias{ask_harvest_pat}
\title{Ask user for Harvest pat}
\usage{
ask_harvest_pat(msg = "Harvest API Bearer Token")
}
\arguments{
\item{msg}{message in the Rstudio prompt}
}
\description{
Ask user for Harvest pat
}
