% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{called_from}
\alias{called_from}
\alias{is_knitting}
\title{Test if a function was called from others.}
\usage{
called_from(..., FUNS = list(...))

is_knitting()
}
\arguments{
\item{...}{functions to pass}

\item{FUNS}{functions as a list}
}
\description{
Test if a function was called from others.
}
\section{Functions}{
\itemize{
\item \code{is_knitting}: Determine if the function is called while knitting a document
}}

