% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_format.R
\name{check_date_format}
\alias{check_date_format}
\title{Check date params for Harvest API v2 standards}
\usage{
check_date_format(query = NULL)
}
\description{
Harvest API v2 accepts a few date parameters which have strict requirements. Date parameters must
be in `yyyy-mm-dd` format. `yyyy/mm/dd`, `mm-dd-yyyy`, `yyyy-m-d`,  etc. will result in failure.
This function simply stops the user if they try an incorrect format.
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
