% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_time_entry.R
\name{coerce_time_entries}
\alias{coerce_time_entries}
\title{R-native time_entry object}
\usage{
coerce_time_entries(time_entries = NULL)
}
\arguments{
\item{time_entry}{a list; *required*; a list (can be partial) of the
time_entry attributes; defaults to NULL}
}
\value{
a list;
}
\description{
Coerces the Harvest time_entry object to R-native data types
}
\details{
Coerces the data types of the elements of the time_entry object from
  Harvest. Coercions are based on the Harvest API docs
  \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/}
   as follows (Harvest Data Type > R Data type): integer > integer, date >
  ___, object > list, decimal > numeric, string > character, boolean >
  logical, datetime > datetime, time > ---

  This function only does coercion for the time_entry object - other objects
  in the time_entries payload are coerced by their respective functions.

  Functions similar to this may eventually be available for other API calls.
}
\examples{
# See get_time_entries.R

}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
