% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_users.R
\name{coerce_users}
\alias{coerce_users}
\title{Harvest User Object with Data Type Conversion}
\usage{
coerce_users(users = NULL)
}
\arguments{
\item{users}{a list; *required*; a list of users with a full or partial list
of the user attributes; defaults to NULL}
}
\value{
a list;
}
\description{
Coerces the Harvest user object to R-native data types
}
\details{
Coerces the data types of the elements of the user object from
  Harvest. Coercions are based on the Harvest API docs
  \href{https://help.getharvest.com/api-v2/users-api/users/users/}
   as follows (Harvest Data Type > R Data type): integer > integer, date >
  Date, object > list, decimal > numeric, string > character, boolean >
  logical, datetime > datetime, time > ---
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
