% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{create_a_time_entry_via_start_and_end_time}
\alias{create_a_time_entry_via_start_and_end_time}
\title{Create a time entry via start and end time}
\usage{
create_a_time_entry_via_start_and_end_time(
  user_id = NULL,
  project_id,
  task_id,
  spent_date,
  started_time = NULL,
  ended_time = NULL,
  notes = NULL,
  external_reference = NULL
)
}
\arguments{
\item{user_id}{\code{(integer)} \strong{optional} The ID of the user to associate with the time entry. Defaults to the currently authenticated user’s ID.}

\item{project_id}{\code{(integer)} \strong{required} The ID of the project to associate with the time entry.}

\item{task_id}{\code{(integer)} \strong{required} The ID of the task to associate with the time entry.}

\item{spent_date}{\code{(date)} \strong{required} The ISO 8601 formatted date the time entry was spent.}

\item{started_time}{\code{(time)} \strong{optional} The time the entry started. Defaults to the current time. Example: “8:00am”.}

\item{ended_time}{\code{(time)} \strong{optional} The time the entry ended. If provided, is_running will be set to false. If not provided, is_running will be set to true.}

\item{notes}{\code{(string)} \strong{optional} Any notes to be associated with the time entry.}

\item{external_reference}{\code{(object)} \strong{optional} An object containing the id, group_id, account_id, and permalink of the external reference.}
}
\description{
Creates a new time entry object. Returns a time entry object and a 201 Created response code if the call succeeded.
You should only use this method to create time entries when your account is configured to track time via start and end time. You can verify this by visiting the Settings page in your Harvest account or by checking if wants_timestamp_timers is true in the Company API.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#create-a-time-entry-via-start-and-end-time}{Create a time entry via start and end time}
}
