% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_harvest_creds}
\alias{create_harvest_creds}
\title{Create or reset the Harvest account ID & pat on the keyring to use as Harvest credentials in the future}
\usage{
create_harvest_creds(keyring = NULL)
}
\arguments{
\item{keyring}{keyring to retreive creds from, use the keyring set in create_harvest_creds (default value of NULL is valid)}
}
\description{
Creates a keyring record for the Harvest account ID & personal access token to use with set_harvest_creds() in future R sessions. New entries replace old credentials.
}
\details{
\lifecycle{experimental}
}
