% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{delete_a_time_entry_s_external_reference}
\alias{delete_a_time_entry_s_external_reference}
\title{Delete a time entry’s external reference}
\usage{
delete_a_time_entry_s_external_reference(TIME_ENTRY_ID)
}
\description{
Delete a time entry’s external reference. Returns a 200 OK response code if the call succeeded.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#delete-a-time-entrys-external-reference}{Delete a time entry’s external reference}
}
