% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_harvest_account}
\alias{get_harvest_account}
\title{Get Harvest account ID from system environment variables}
\usage{
get_harvest_account()
}
\description{
Retrieve the harvestR account ID from system environment variables under the default alias HARVEST_ACCOUNT_ID
}
\details{
\lifecycle{experimental}
}
