% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_harvestR.r
\name{get_harvestR}
\alias{get_harvestR}
\title{Harvest API v2 get request wrapper}
\usage{
get_harvestR(table = NULL, user = NULL, key = NULL, email = "",
  query = NULL)
}
\arguments{
\item{table}{Harvest API table or tables to call. Some tables have not been included, but may be added in the future. Refer to \href{https://help.getharvest.com/api-v2}{Harvest APIv2}. Possible values include any combination of the following strings in a vector:  (\code{"clients"} or \code{"invoices"} or \code{"estimates"} or \code{"expenses"} or \code{"tasks"} or \code{"time_entries"} or \code{"projects"} or \code{"roles"} or \code{"users"} or \code{"user_project_assignments"} or \code{"project_task_assignments"} or \code{"project_user_assignments"} or \code{"all"})}

\item{user}{Harvest account number for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{key}{Harvest API key for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{email}{Optional argument, requester email address for the "From" header in the get request.}

\item{query}{Optional argument, API query parameters to be provided in a list. Refer to \href{https://help.getharvest.com/api-v2}{Harvest APIv2} for acceptable parameters for each table. A few examples include:  (\code{from = "2018-1-1"} or \code{to = "2018-3-31"} or \code{project_id = "1234"} or \code{client_id = "1234"} or \code{user_id = "1234"})}
}
\value{
A data frame for each Harvest table with the name of the Harvest table e.g. 'clients'.
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and returns Harvest table(s) with their respective names from the API documentation.
}
\examples{

user_id <- 'your_user_id'
api_key <- paste0('Bearer ','you_api_key')
get_harvestR(table = c('clients','users','projects') , user = user_id,key = api_key)
get_harvestR(table = c('time_entries') , user = user_id, key = api_key, email = 'your_email',query = list(from = '2018-1-1', to = '2018-3-31'))

}
\references{
\url{https://id.getharvest.com/developers}, \url{https://help.getharvest.com/api-v2}, \url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
