% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Harvest API v2 get wrapper}
\usage{
get_request(
  url = NULL,
  user = NULL,
  key = NULL,
  email = NULL,
  verbose = F,
  auto_retry = F,
  ...
)
}
\arguments{
\item{url}{tbd}

\item{user}{Harvest account number for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{key}{Harvest API key for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{email}{Optional argument, requester email address for the "From" header in the get request.}

\item{verbose}{logical; passed to httr; A verbose connection provides much more information about the flow of information between the client and server. \href{https://github.com/r-lib/httr}{httr documentation}}

\item{auto_retry}{logical;}

\item{...}{tbd}
}
\value{
named list
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and safely returns the httr result as result or error.
}
\examples{


}
\references{
\url{https://id.getharvest.com/developers},
\url{https://help.getharvest.com/api-v2},
\url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
