% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requests.R
\name{get_requests}
\alias{get_requests}
\title{Harvest API v2 get wrapper}
\usage{
get_requests(
  urls = NULL,
  group_name = NULL,
  user = user,
  key = key,
  email = email,
  verbose = verbose,
  auto_retry = auto_retry,
  plan = plan,
  ...
)
}
\arguments{
\item{url}{tbd}
}
\value{
named list
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and safely returns the httr result as result or error.
}
\examples{


}
\references{
\url{https://id.getharvest.com/developers},
\url{https://help.getharvest.com/api-v2},
\url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
