% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_requests_lim}
\alias{get_requests_lim}
\title{Execute multiple get_request calls with rate limiting}
\usage{
get_requests_lim(urls, table, user, key, ...)
}
\description{
Execute multiple get_request calls with rate limiting
}
