% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_entries.R
\name{get_time_entries}
\alias{get_time_entries}
\title{Get normalized Harvest time_entries}
\usage{
get_time_entries(...)
}
\arguments{
\item{...}{additional parameters passed to get_harvestR()}
}
\value{
a list; json object converted to R list
}
\description{
Requests time_entries from Harvest and then normalizes them
}
\details{
This function requests time_entries from get_harvestR(), then
  normalizes them. Right now, normalization only includes data type coercion
  to native R data types and structures.
}
\examples{
\dontrun{
get_time_entries(user = user_id,
                 key = api_key,
                 email = 'your_email',
                 query = list(
                   from = '2018-01-01',
                   to = '2018-03-31')
                 )
}

}
\references{
\url{https://help.getharvest.com/api-v2},
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
