% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_user_agent}
\alias{get_user_agent}
\title{Get User-Agent Environmental Variable}
\usage{
get_user_agent()
}
\description{
Retrieve the harvestR HARVESTR_USER_AGENT variable
}
\details{
\lifecycle{experimental}
}
