% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttr}
\alias{getAttr}
\title{Retrieve an attribute or a default if not present.}
\usage{
getAttr(object, name, default = NULL)
}
\arguments{
\item{object}{An R Object.}

\item{name}{Name of the Attribute}

\item{default}{The value if the attribute is not set or NULL}
}
\description{
Behaves similar to \code{\link{getOption}}, but is a simple wrapper
for \code{\link{attr}}.
}

