% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{harvest_pat}
\alias{harvest_pat}
\title{Retrieve Harvest Personal Access Token}
\usage{
harvest_pat()
}
\value{
A string of the Harvest Personal Access Token
}
\description{
Looks in env var HARVEST_PAT
}
