% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{Interactive}
\alias{Interactive}
\title{Smarter interactive test}
\usage{
Interactive(exclude.calls = getOption("harvestr::Interactive::exclude"))
}
\arguments{
\item{exclude.calls}{functions to pass to \code{\link{called_from}}}
}
\description{
This is a smarter version of \code{\link{interactive}},
but also excludes cases inside knit or in startup 
\code{\link{.First}}, or others specified in dots.
You can also specify functions to exclude in the option
\code{harvestr::Interactive::exclude}
}

