% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is_seeded}
\alias{is_seeded}
\title{Check if an object or list of objects has seed attributes}
\usage{
is_seeded(x)
}
\arguments{
\item{x}{an object or list to check}
}
\description{
Check if an object or list of objects has seed attributes
}

