\name{noattr}
\alias{noattr}
\title{Strip attributes}
\usage{
  noattr(x)
}
\arguments{
  \item{x,}{any object}
}
\description{
  Strip attributes
}
\seealso{
  Other harvest: \code{\link{farm}}, \code{\link{gather}},
  \code{\link{graft}}, \code{\link{harvest}},
  \code{\link{plant}}, \code{\link{sprout}}
}

