% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paginate}
\alias{paginate}
\title{Fetch extra pages (if present)}
\usage{
paginate(resp, .args)
}
\arguments{
\item{resp}{\code{(response)} The \link[httr]{response} object}

\item{.args}{\code{(list)} The \code{.args} provided to \link[httr]{RETRY}}
}
\description{
Some endpoints return pages. This function fetches and collates them.
}
