% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{restart_a_stopped_time_entry}
\alias{restart_a_stopped_time_entry}
\title{Restart a stopped time entry}
\usage{
restart_a_stopped_time_entry(TIME_ENTRY_ID)
}
\description{
Restarting a time entry is only possible if it isn’t currently running. Returns a 200 OK response code if the call succeeded.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#restart-a-stopped-time-entry}{Restart a stopped time entry}
}
