% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{retrieve_a_time_entry}
\alias{retrieve_a_time_entry}
\title{Retrieve a time entry}
\usage{
retrieve_a_time_entry(
  TIME_ENTRY_ID,
  project_id,
  task_id,
  spent_date,
  user_id = NULL,
  hours = NULL,
  notes = NULL,
  external_reference = NULL
)
}
\arguments{
\item{TIME_ENTRY_ID}{\code{(integer)} \strong{required} The ID of the time entry with which the call will be associated.}

\item{project_id}{\code{(integer)} \strong{required} The ID of the project to associate with the time entry.}

\item{task_id}{\code{(integer)} \strong{required} The ID of the task to associate with the time entry.}

\item{spent_date}{\code{(date)} \strong{required} The ISO 8601 formatted date the time entry was spent.}

\item{user_id}{\code{(integer)} \strong{optional} The ID of the user to associate with the time entry. Defaults to the currently authenticated user’s ID.}

\item{hours}{\code{(decimal)} \strong{optional} The current amount of time tracked. If provided, the time entry will be created with the specified hours and is_running will be set to false. If not provided, hours will be set to 0.0 and is_running will be set to true.}

\item{notes}{\code{(string)} \strong{optional} Any notes to be associated with the time entry.}

\item{external_reference}{\code{(object)} \strong{optional} An object containing the id, group_id, account_id, and permalink of the external reference.}
}
\description{
Retrieves the time entry with the given ID. Returns a time entry object and a 200 OK response code if a valid identifier was provided.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#retrieve-a-time-entry}{Retrieve a time entry}
}
