% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_time_entries}
\alias{round_time_entries}
\title{Round time entries to the hour decimal}
\usage{
round_time_entries(
  entries = list_all_time_entries(from = lubridate::floor_date(Sys.Date() - 2, "week")),
  round_hours = 0.25
)
}
\arguments{
\item{entries}{\code{(tbl)} of the entries to round}

\item{round_hours}{\code{(numeric)} the decimal hours to round to (IE 15 min = .25)}
}
\value{
\code{(tbl)} of changed entries with a \code{prev_hours} column with the previous amount of hours and a \code{difference} column with the total hours added
}
\description{
Round time entries to the hour decimal
}
