% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_harvest_creds}
\alias{set_harvest_creds}
\title{Set the Harvest account ID & pat as system environment variables for the R session}
\usage{
set_harvest_creds(keyring = NULL)
}
\arguments{
\item{keyring}{keyring to retreive creds from, use the keyring set in create_harvest_creds (default value of NULL is valid)}
}
\description{
Retrieve the Harvest account ID & personal access token from the keyring and set them as environment variables for the R session
}
\details{
\lifecycle{experimental}
}
\seealso{
[keyring::key_get()] which this function wraps.
}
