% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_harvestR}
\alias{setup_harvestR}
\title{Setup harvestR}
\usage{
setup_harvestR(user_agent, overwrite = TRUE)
}
\arguments{
\item{user_agent}{\code{(chr)} An identifier and email address to set as your User-Agent. See link in description for details.}

\item{overwrite}{\code{(log)} Whether to overwrite the existing value for HARVESTR_USER_AGENT in the user-level \emph{.Renviron} file.}
}
\value{
Nothing. Used for side-affects.
}
\description{
The Harvest API requires the account ID,API key and a User Agent header for each request. See the \href{https://help.getharvest.com/api-v2/introduction/overview/general/#api-requests}{API Overview} for details.
}
\seealso{
create_harvest_creds, set_harvest_creds
}
