% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{stop_a_running_time_entry}
\alias{stop_a_running_time_entry}
\title{Stop a running time entry}
\usage{
stop_a_running_time_entry(TIME_ENTRY_ID)
}
\description{
Stopping a time entry is only possible if it’s currently running. Returns a 200 OK response code if the call succeeded.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#stop-a-running-time-entry}{Stop a running time entry}
}
