% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheets.R
\name{the_time_entry_object}
\alias{the_time_entry_object}
\title{The time entry object}
\description{
Returns a list of your time entries. The time entries are returned sorted by spent_date date. At this time, the sort option can’t be customized.
The response contains an object with a time_entries property that contains an array of up to per_page time entries. Each entry in the array is a separate time entry object. If no more time entries are available, the resulting array will be empty. Several additional pagination properties are included in the response to simplify paginating your time entries.
From \href{https://help.getharvest.com/api-v2/timesheets-api/timesheets/time-entries/#the-time-entry-object}{The time entry object}
\itemize{
\item{id}{\code{(integer)} Unique ID for the time entry.}
\item{spent_date}{\code{(date)} Date of the time entry.}
\item{user}{\code{(object)} An object containing the id and name of the associated user.}
\item{user_assignment}{\code{(object)} A user assignment object of the associated user.}
\item{client}{\code{(object)} An object containing the id and name of the associated client.}
\item{project}{\code{(object)} An object containing the id and name of the associated project.}
\item{task}{\code{(object)} An object containing the id and name of the associated task.}
\item{task_assignment}{\code{(object)} A task assignment object of the associated task.}
\item{external_reference}{\code{(object)} An object containing the id, group_id, account_id, permalink, service, and service_icon_url of the associated external reference.}
\item{invoice}{\code{(object)} Once the time entry has been invoiced, this field will include the associated invoice’s id and number.}
\item{hours}{\code{(decimal)} Number of (decimal time) hours tracked in this time entry.}
\item{hours_without_timer}{\code{(decimal)} Number of (decimal time) hours already tracked in this time entry, before the timer was last started.}
\item{rounded_hours}{\code{(decimal)} Number of (decimal time) hours tracked in this time entry used in summary reports and invoices. This value is rounded according to the Time Rounding setting in your Preferences.}
\item{notes}{\code{(string)} Notes attached to the time entry.}
\item{is_locked}{\code{(boolean)} Whether or not the time entry has been locked.}
\item{locked_reason}{\code{(string)} Why the time entry has been locked.}
\item{is_closed}{\code{(boolean)} Whether or not the time entry has been approved via Timesheet Approval.}
\item{is_billed}{\code{(boolean)} Whether or not the time entry has been marked as invoiced.}
\item{timer_started_at}{\code{(datetime)} Date and time the timer was started (if tracking by duration). Use the ISO 8601 Format.}
\item{started_time}{\code{(time)} Time the time entry was started (if tracking by start/end times).}
\item{ended_time}{\code{(time)} Time the time entry was ended (if tracking by start/end times).}
\item{is_running}{\code{(boolean)} Whether or not the time entry is currently running.}
\item{billable}{\code{(boolean)} Whether or not the time entry is billable.}
\item{budgeted}{\code{(boolean)} Whether or not the time entry counts towards the project budget.}
\item{billable_rate}{\code{(decimal)} The billable rate for the time entry.}
\item{cost_rate}{\code{(decimal)} The cost rate for the time entry.}
\item{created_at}{\code{(datetime)} Date and time the time entry was created. Use the ISO 8601 Format.}
\item{updated_at}{\code{(datetime)} Date and time the time entry was last updated. Use the ISO 8601 Format.}
}
}
