% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_entry_tag_sums}
\alias{time_entry_tag_sums}
\title{Summarize by feature tags}
\usage{
time_entry_tag_sums(entries)
}
\arguments{
\item{entries}{}
}
\value{
\code{list}
}
\description{
Summarize by feature tags
}
