% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url_groups.R
\name{build_url_groups}
\alias{build_url_groups}
\title{Harvest API v2 get wrapper}
\usage{
build_url_groups(urls = NULL, verbose = verbose)
}
\arguments{
\item{url}{tbd}
}
\value{
named list
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and safely returns the httr result as result or error.
}
\examples{


}
\references{
\url{https://id.getharvest.com/developers},
\url{https://help.getharvest.com/api-v2},
\url{https://github.com/r-lib/httr}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
