% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather}
\alias{gather}
\title{Gather independent seeds.}
\usage{
gather(x, seed = get.seed(), ..., .starting = F)
}
\arguments{
\item{x}{number of seeds, or an object with seeds to gather}

\item{seed}{a seed to use to set the seed, must be compatible with "L'Ecuyer-CMRG"}

\item{...}{passed on}

\item{.starting}{if TRUE starting seeds with be gathered rather than ending seeds.}
}
\description{
Collect seeds representing independent random number streams.
These seeds can them be used in \code{\link{farm}} or \code{\link{plant}}.
}
\seealso{
\link{RNG}

Other harvest: \code{\link{farm}}, \code{\link{harvest}},
  \code{\link{plant}}, \code{\link{sprout}}
}

