% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Harvest API v2 get request wrapper}
\usage{
get_request(url = NULL, user = NULL, key = NULL, ...)
}
\arguments{
\item{url}{Character url of the get request.}

\item{user}{Character account number (e.g. user = '12345') for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{key}{Character key (e.g. key = 'Bearer <secret key>') for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{...}{Additional arguments, potentially passed to other functions - see details in \link[harvestR]{get_table}}
}
\value{
contents of the API response
}
\description{
Submits get requests to the Harvest API v2 with retry functionality.
Errors are messaged to the execution environment and set to NULL allowing user-facing function returns to provide data frame output.
}
\references{
Set up your \href{https://id.getharvest.com/developers}{API key}
Read the Harvest API v2 \href{https://help.getharvest.com/api-v2}{documentation}
Read the httr \href{https://github.com/r-lib/httr}{documentation}
Refer to curl \href{https://github.com/curl/curl/blob/master/include/curl/curl.h}{documentation} for more http_version options
}
