% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_requests}
\alias{get_requests}
\title{Execute multiple get_request calls}
\usage{
get_requests(urls, table, user, key, ...)
}
\arguments{
\item{urls}{Named list of character vectors of urls in groups from build_url_groups function.}

\item{table}{Character table name to be submitted to the Harvest API v2. Refer to \href{https://help.getharvest.com/api-v2}{Harvest API v2}.}

\item{user}{Character account number (e.g. user = '12345') for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{key}{Character key (e.g. key = 'Bearer <secret key>') for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{...}{Additional arguments, potentially passed to other functions - see details in \link[harvestR]{get_table}}
}
\value{
named list of contents of API responses for specified group of URLs
}
\description{
Submits multiple get requests to the Harvest API v2. Supports parallel execution through furrr & the future package.
}
\references{
\url{https://github.com/HenrikBengtsson/future} \cr
\url{https://davisvaughan.github.io/furrr} \cr
\url{https://rstudio.com/resources/rstudioconf-2020/parallel-computing-with-r-using-foreach-future-and-other-packages} \cr
\url{https://help.getharvest.com/api-v2/introduction/overview/general}
}
\seealso{
[future::plan()] for details regarding how to implement parallel api calls
}
