% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Harvest API v2 get request wrapper}
\usage{
get_table(
  table = NULL,
  user = harvestR::get_harvest_account(),
  key = harvestR::get_harvest_pat(),
  email = "",
  query = NULL,
  ...,
  verbose = lifecycle::deprecated(),
  strategy = lifecycle::deprecated()
)
}
\arguments{
\item{table}{Character table name to be submitted to the Harvest API v2. Refer to \href{https://help.getharvest.com/api-v2}{Harvest API v2}.}

\item{user}{Character account number (e.g. user = '12345') for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}. Defaults to `get_harvest_account()`}

\item{key}{Character key (e.g. key = 'Bearer <secret key>') for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}. Defaults to `get_harvest_pat()`}

\item{email}{Character email address for API call. (Optional)}

\item{...}{Additional arguments, some are forwarded to other functions - see details.}
}
\value{
dataframe with contents of requested Harvest table
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and returns the Harvest table.
}
\details{
\lifecycle{maturing}


Description of ... parameters
\itemize{
   \item \code{auto_retry} Logical; control whether to retry GET requests upon failure. TRUE results 2 retries (3 attempts total). FALSE results in no retries (1 attempt total).
   \item \code{quiet} Logical; controls messaging. TRUE will result in displaying only errors in the console. FALSE will results in displaying messages, warnings, and errors. Depending on execution strategy, messages from httr::GET may not be surfaced to the console due to multiple R sessions running and stderror relaying issues. See \href{https://cran.r-project.org/web/packages/future/vignettes/future-2-output.html}{Future Package Output} for me details.
   \item \code{plan_options} List; list of parameters passed to \code{future::plan}. The strategy parameter is required to implement parallel requests, see \href{https://cran.r-project.org/web/packages/future/vignettes/future-1-overview.html}{Future Package Overview} for more details.
   }
}
\examples{
\dontrun{
create_harvest_creds()
set_harvest_creds()
get_table(table = 'projects')
get_table(table = 'time_entries',
          email = 'your_email',
          query = list(from = '2018-01-01', to = '2018-12-31'),
          plan_options = list(strategy = "multicore"),
          auto_retry = T,
          quiet = T)
}

}
\references{
\href{https://help.getharvest.com/api-v2}{Harvest API v2}
\href{https://id.getharvest.com/developers}{Harvest Developers Page}
\href{https://cran.r-project.org/web/packages/future/vignettes/future-1-overview.html}{Future Pacakge Overview}
\href{https://cran.r-project.org/web/packages/future/vignettes/future-2-output.html}{Future Package Output}
}
\seealso{
`future::plan()` for more detail regarding how to implement parallel api calls
}
