% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Harvest API v2 get request wrapper}
\usage{
get_table(
  table = NULL,
  user = NULL,
  key = NULL,
  email = "",
  query = NULL,
  auto_retry = FALSE,
  strategy = "sequential",
  verbose = FALSE
)
}
\arguments{
\item{table}{Character table name to be submitted to the Harvest API v2. Refer to \href{https://help.getharvest.com/api-v2}{Harvest APIv2}.}

\item{user}{Character account number (e.g. user = '12345') for authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{key}{Character key (e.g. key = 'Bearer <secret key>') for API authentication. Register at \href{https://id.getharvest.com/developers}{Harvest Developers}.}

\item{email}{Character email address for API call. (Optional)}

\item{strategy}{Character vector passed to future::plan strategy parameter. Examples include 'sequential', 'multiprocess', & 'multicore'.}

\item{verbose}{Logical scalar. Should the function provide verbose messaging back on each step?}
}
\description{
Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and returns the Harvest table.
}
\examples{

user_id <- 'your_user_id'
api_key <- paste0('Bearer ','you_api_key')
get_table(table = 'projects', user = user_id,key = api_key)
get_table(table = 'time_entries' , user = user_id, key = api_key, email = 'your_email',query = list(from = '2018-01-01', to = '2018-03-31'))

}
\references{
\url{https://id.getharvest.com/developers}
\url{https://help.getharvest.com/api-v2}
}
\author{
Mark Druffel, \email{mdruffel@propellerpdx.com}
}
