% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{harvest}
\alias{harvest}
\title{Harvest the results.}
\usage{
harvest(.list, fun, ..., time = getOption("harvestr.time", defaults$time()),
  .parallel = getOption("harvestr.parallel", defaults$parallel()),
  .progress = getOption("harvestr.progress", defaults$progress()))
}
\arguments{
\item{.list}{a list of \code{data.frame}s  See details.}

\item{fun}{a function to apply}

\item{...}{passed to \code{fun}}

\item{time}{should results be timed?}

\item{.parallel}{should the computations be run in parallel?}

\item{.progress}{Show a progress bar?}
}
\description{
Harvest the results.
}
\details{
harvest is functionaly equivalant to llply, but takes on additional capability when used
with the other functions from this package.  When an object comes from \code{\link{withseed}}
the ending seed is extacted and used to continue evaluation.
}
\seealso{
Other harvest: \code{\link{farm}}, \code{\link{gather}},
  \code{\link{plant}}, \code{\link{sprout}}
}

