\name{reap}
\alias{reap}
\title{call an object continuing the random number stream.}
\usage{
  reap(x, fun, ...)
}
\arguments{
  \item{x}{an object}

  \item{fun}{a function to call on object}

  \item{...}{passed onto function}
}
\description{
  call an object continuing the random number stream.
}
\seealso{
  \code{\link{withpseed}}, \code{\link{harvest}}, and
  \code{\link{with}}
}

